# backend/openapi_generated.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AGRO",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/dispatch-entries/": {
      "post": {
        "tags": [
          "Dispatch Entries"
        ],
        "summary": "Create Route",
        "description": "Create a new dispatch entry.\n\nArgs:\n    entry (DispatchEntryCreate): Dispatch entry data.\n    db (Session): Database session dependency.\n\nReturns:\n    DispatchEntryRead: The created dispatch entry.",
        "operationId": "create_route_v1_dispatch_entries__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DispatchEntryCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DispatchEntryRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Dispatch Entries"
        ],
        "summary": "Read All",
        "description": "Retrieve all dispatch entries with optional filters.\n\nArgs:\n    skip (int): Number of records to skip.\n    limit (int): Maximum number of records to return.\n    dispatch_date (Optional[date]): Filter by dispatch date.\n    mart_name (Optional[str]): Filter by mart name.\n    db (Session): Database session dependency.\n\nReturns:\n    List[DispatchEntryRead]: List of dispatch entries.",
        "operationId": "read_all_v1_dispatch_entries__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "dispatch_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dispatch Date"
            }
          },
          {
            "name": "mart_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mart Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DispatchEntryRead"
                  },
                  "title": "Response Read All V1 Dispatch Entries  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/dispatch-entries/from-order": {
      "post": {
        "tags": [
          "Dispatch Entries"
        ],
        "summary": "Dispatch From Order",
        "description": "Create multiple dispatch entries from an order.\n\nArgs:\n    entry (DispatchEntryMultiCreate): Order-based dispatch data.\n    db (Session): Database session dependency.\n\nReturns:\n    List[DispatchEntryRead]: List of created dispatch entries.\n\nRaises:\n    AppException: If creation fails (400).",
        "operationId": "dispatch_from_order_v1_dispatch_entries_from_order_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DispatchEntryMultiCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DispatchEntryRead"
                  },
                  "type": "array",
                  "title": "Response Dispatch From Order V1 Dispatch Entries From Order Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/v1/dispatch-entries/{id}": {
      "get": {
        "tags": [
          "Dispatch Entries"
        ],
        "summary": "Read One",
        "description": "Retrieve a single dispatch entry by ID.\n\nArgs:\n    id (int): Dispatch entry ID.\n    db (Session): Database session dependency.\n\nReturns:\n    DispatchEntryRead: The dispatch entry.\n\nRaises:\n    AppException: If the entry is not found (404).",
        "operationId": "read_one_v1_dispatch_entries__id__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DispatchEntryRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Dispatch Entries"
        ],
        "summary": "Update Route",
        "description": "Update a dispatch entry by ID.\n\nArgs:\n    id (int): Dispatch entry ID.\n    entry (DispatchEntryUpdate): Update data.\n    db (Session): Database session dependency.\n\nReturns:\n    DispatchEntryRead: The updated dispatch entry.\n\nRaises:\n    AppException: If the entry is not found (404).",
        "operationId": "update_route_v1_dispatch_entries__id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DispatchEntryUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DispatchEntryRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Dispatch Entries"
        ],
        "summary": "Delete Route",
        "description": "Delete a dispatch entry by ID.\n\nArgs:\n    id (int): Dispatch entry ID.\n    db (Session): Database session dependency.\n\nRaises:\n    AppException: If the entry is not found (404).",
        "operationId": "delete_route_v1_dispatch_entries__id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/dispatch/dispatch-entries/": {
      "post": {
        "tags": [
          "Dispatch Entry",
          "Dispatch Entries"
        ],
        "summary": "Create Route",
        "description": "Create a new dispatch entry.\n\nArgs:\n    entry (DispatchEntryCreate): Dispatch entry data.\n    db (Session): Database session dependency.\n\nReturns:\n    DispatchEntryRead: The created dispatch entry.",
        "operationId": "create_route_v1_dispatch_dispatch_entries__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DispatchEntryCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DispatchEntryRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Dispatch Entry",
          "Dispatch Entries"
        ],
        "summary": "Read All",
        "description": "Retrieve all dispatch entries with optional filters.\n\nArgs:\n    skip (int): Number of records to skip.\n    limit (int): Maximum number of records to return.\n    dispatch_date (Optional[date]): Filter by dispatch date.\n    mart_name (Optional[str]): Filter by mart name.\n    db (Session): Database session dependency.\n\nReturns:\n    List[DispatchEntryRead]: List of dispatch entries.",
        "operationId": "read_all_v1_dispatch_dispatch_entries__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "dispatch_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dispatch Date"
            }
          },
          {
            "name": "mart_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mart Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DispatchEntryRead"
                  },
                  "title": "Response Read All V1 Dispatch Dispatch Entries  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchDispatchInput": {
        "properties": {
          "batch_id": {
            "type": "integer",
            "title": "Batch Id"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          }
        },
        "type": "object",
        "required": [
          "batch_id",
          "quantity"
        ],
        "title": "BatchDispatchInput"
      },
      "BatchRead": {
        "properties": {
          "item_id": {
            "type": "integer",
            "title": "Item Id"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "received_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Received At"
          },
          "expiry_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiry Date"
          },
          "remarks": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remarks"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "item_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Name"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "updated_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated By"
          }
        },
        "type": "object",
        "required": [
          "item_id",
          "quantity",
          "unit",
          "id"
        ],
        "title": "BatchRead"
      },
...
      "DispatchEntryCreate": {
        "properties": {
          "item_id": {
            "type": "integer",
            "title": "Item Id"
          },
          "batch_id": {
            "type": "integer",
            "title": "Batch Id"
          },
          "mart_name": {
            "type": "string",
            "title": "Mart Name"
          },
          "dispatch_date": {
            "type": "string",
            "format": "date",
            "title": "Dispatch Date"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "remarks": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remarks"
          }
        },
        "type": "object",
        "required": [
          "item_id",
          "batch_id",
          "mart_name",
          "dispatch_date",
          "quantity",
          "unit"
        ],
        "title": "DispatchEntryCreate"
      },
      "DispatchEntryMultiCreate": {
        "properties": {
          "item_id": {
            "type": "integer",
            "title": "Item Id"
          },
          "mart_name": {
            "type": "string",
            "title": "Mart Name"
          },
          "dispatch_date": {
            "type": "string",
            "format": "date",
            "title": "Dispatch Date"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "remarks": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remarks"
          },
          "batches": {
            "items": {
              "$ref": "#/components/schemas/BatchDispatchInput"
            },
            "type": "array",
            "title": "Batches"
          }
        },
        "type": "object",
        "required": [
          "item_id",
          "mart_name",
          "dispatch_date",
          "unit",
          "batches"
        ],
        "title": "DispatchEntryMultiCreate"
      },
      "DispatchEntryRead": {
        "properties": {
          "item_id": {
            "type": "integer",
            "title": "Item Id"
          },
          "batch_id": {
            "type": "integer",
            "title": "Batch Id"
          },
          "mart_name": {
            "type": "string",
            "title": "Mart Name"
          },
          "dispatch_date": {
            "type": "string",
            "format": "date",
            "title": "Dispatch Date"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "remarks": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remarks"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "ISO 8601 format"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "ISO 8601 format"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "updated_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated By"
          },
          "batch": {
            "$ref": "#/components/schemas/BatchRead"
          }
        },
        "type": "object",
        "required": [
          "item_id",
          "batch_id",
          "mart_name",
          "dispatch_date",
          "quantity",
          "unit",
          "id",
          "created_at",
          "updated_at",
          "created_by",
          "updated_by",
          "batch"
        ],
        "title": "DispatchEntryRead"
      },
      "DispatchEntryUpdate": {
        "properties": {
          "mart_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mart Name"
          },
          "dispatch_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dispatch Date"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          }
        },
        "type": "object",
        "title": "DispatchEntryUpdate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/token"
          }
        }
      }
    }
  }
}